ocr/requirements.txt
anyio==3.7.1 #anyio==4.3.0
appnope==0.1.3
astor==0.8.1
asttokens==2.4.0
attrdict==2.0.1
Babel==2.14.0
backcall==0.2.0
bce-python-sdk==0.9.5
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
contourpy==1.2.0
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.9
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fire==0.6.0
Flask==3.0.2
flask-babel==4.0.0
fonttools==4.50.0
future==1.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imageio==2.34.0
imgaug==0.4.0
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.3
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
paddleocr==2.7.0.3
paddlepaddle==2.6.1
pandas==2.2.1
parso==0.8.3
pdf2docx==0.5.8
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.10.0
premailer==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.2 #protobuf==4.21.9
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyclipper==1.3.0.post5
pycryptodome==3.20.0
Pygments==2.13.0
PyMuPDF==1.20.2
pyparsing==3.1.2
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.6.2
rarfile==4.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
termcolor==2.4.0
tifffile==2024.2.12
tornado==6.3.3
tqdm==4.66.2
traitlets==5.10.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
visualdl==2.5.3
wcwidth==0.2.6
Werkzeug==3.0.1
